[Fix]: Moved unit testing code to debug.py, restored test.py.
test.py should only contain code related to model evaluation.
